requirements.txt
anyio==3.7.0
APScheduler==3.6.3
cachetools==4.2.2
certifi==2023.5.7
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
python-telegram-bot==20.3
pytz==2023.3
six==1.16.0
sniffio==1.3.0
tornado==6.3.2
tzlocal==5.0.1
peewee==3.14.0